// frontend/types/Customer.ts
export interface Customer {
  id: number
  first_name: string
  last_name: string
  address: string
  phone: string
  company_name: string
  tax_id: string
  email: string
  status: string
  credit_balance: number
  document_type: string
  document_number: string
  city: string
  department: string
  country: string
}